Add docs about nuget and PackageReference format
Trying to add the SimpleRemote Client nuget package to a .NET
Framework project results in a "could not resolve dependency" error.
This is due to a bug in the packages.config package management format,
used by default in .NET Framework. To work around this, consumers
of the library need to use the PackageReference package management
format.

This change adds a note in the HTML docs on how to do this.