Fix issue in RpcClient where tasks completed twice
There was an issue when running the units tests where the timers
in RunJob in RpcClient could fire (and attempt to transition the
task completion source's state) after the test had completed.

To resolve, the timers and callbacks now use the TrySet... methods
when interacting with the task completion source.