Fix issue with extra DLLs in client directory.
If there were extra DLLs in the same location as
SimpleRemoteClient, the redirector would attempt to load them,
potentially triggering DLL load exceptions (especially if they
were unmanaged). Update the redirector to load DLLs it can, and
to simply ignore ones where it cannot get the AssemblyName from
the file.