Version bump to 1.2.0
Update versions in csproj and in ISS files.

Also adds a new python script, SetVersion, to help with easily
updating the appropriate files for versioning.

Also corrects a spelling error in the iss file were the MyAppName
field had a spelling error.

Also adds instructions for how to run as a service to the tutorial
documentation.