Overhauls job streaming system and adds streaming tests.
Significantly improves streaming functionality by ensuring all
streamed progress is also written to a fallback file. It also
moves streaming operations to a separate thread to ensure we're
not waiting in the ProcessOutput callback.

It also adds several tests for the new streaming system, including
checks for streaming when the client isn't listening, and if the
client abruptly terminates the connection.