Fix Upload bug when using relative paths.
When calling the client's upload function with relative paths, there
was an case in which the file would be uploaded, but it would be
missing the first letter of the filename. This is due to a bug
in our tar library (SharpZipLib) - issue 334.

As a work around, whenever relative paths are fed into
TarFunctions.WriteFileOrDirectoryToStream, they will be prepended
with "./" - this resolves the issue.

Additionally, a new test case has been added to ensure relative paths
work with the upload function, and our version of SharpZipLib has
been upgraded to 1.1.0.