Work around for SharpZipLib issue #334, #337, #338
This commit contains work arounds for issues a user found in the
SimpleRemoteClient, where Upload behaves differently if the file to
upload is located in a child directory of the current working
directory. This is related to an issue in SharpZipLib.

We've filed bugs against SharpZipLib, but, for the interim, these
fixes should prevent the odd behavior.