Replace StringBuilder with ThreadSafeStringBuilder
While incredibly rare, it is theoretically possible for an event handler to try to call StringBuilder.Append while another thread is reading the contents of the StringBuilder. This triggers an exception.

This change introduces a new class, ThreadSafeStringBuilder, that wraps the necessary functionality of StringBuilder while also performing locking to prevent simultaneous accesses.