Add GetJobResultEx/RunJobAsyncEx.
Adds a new method on the server, GetJobResultEx, that returns a json object instead of a simple string. The object currently contains the output (like GetJobResult), and the process' exit code, but can be expanded in later versions to include more information.

Also added a new client method, RunJobAsyncEx that internally uses GetJobResultEx instead of GetJobResult, so it's possible to access the new information from the client.