Initial attempt to fix output race condition.
In some rare cases, users reported that, when the nlog target is
syncronous, the log receives all process output but the thread-safe
string builder is missing lines.

Per API guidance, we've added a WaitForExit call on the process object
before reading the string builder, which should wait for the process'
output streams to flush.

If this fails, we'll need to consider changing how the job system
reads process output (either doing syncronous reads on a background
thread, and/or merging log and thread-safe stringbuilder writes into
a single event listener)