Update NLog to 4.7 and fix BuildAll.ps1
This corrects an error the build script, where, when building the server
app, the dotnet command targeted netcoreapp2.0, when the project targets
2.1, resulting in a build error.

It also upgrades the NLog version in the project to 4.7. The past
release on was on an old, pre-release version and there's no reason not
to upgrade to something more stable.