Update csprojs for nuget submissions.
This commit changes data in some of the csproj files to ensure the
packages meets the requirements for nuget submission.

At this time, we're only going to post the client, common, and
rpc libraries to nuget. If there's demand for the server library,
we'll move that to nuget as well.