Remove async network calls in synchronous code
As a work around to TcpClient not having a timeout option on Connect, we
used ConnectAsync with a call to Wait. However, this is bad practice,
since we're calling async methods from non-async code.

To resolve the issue, we've written an extension on TcpClient that
wraps the BeginConnect/EndConnect methods with a timeout. This allows us
to set a timeout while still keeping the majority of the code simple,
and ensures the call is safe from both async and non-async methods.

Additionally, some previous issues with our TcpClient calls to Connect
were related to Connect throwing exceptions on background threads, and
not being properly handled. The extension method returns true on
successful conection, and false if there was an exception or timeout.

We've also swapped other async network calls with equivalent Begin/End
calls in synchronous code. This prevents deadlocking in certain cases.