Fix failing ClientTests + Update RpcClient Exception
Update the Client tests to look for the correct exception types (in
previous versions of the code, before commit f27fa58, connection
failures threw different types of exceptions).

Additionally - fixed the RpcClient to throw a generic exception (rather
then a timeout exception) if it failed to connect. There are plenty of
cases where it may fail for other reasons, and always throwing a timeout
exception was misleading.