Fix Race Condition in Job Completion Callback
The Job code had a bug where the dispose method manually disposed of the
encapsulated Process object. This created issues when a job was killed,
as we'd dispose of the process before the Exitted event would fire.

As a short term solution, we've commented out the call to
process.Dispose() in the Job's dispose method. This resolves the issue
and callbacks should work again, even after called StopJob.

That said, in the future, we should likely rethink how we're handling
Job object disposal and see if there's a more elegant solution.