Update SimpleRpcServer to log non-exit exceptions
Ensure that the main server loop logs any thrown exceptions when not
shutting down.

Also update the code comments to correctly identify the new type of
exception thrown in .NET 6. In .NET Core 3.1, closing the waiting
listener would generate an Object Disposed Exception, while in .NET 6 is
generates a socket exception (it was incorrectly marked as an Aggregate
Exception earlier).